requirements.txt
numpy==1.26.4
Pillow==9.4.0
pip==24.3.1
pygame==2.5.2
pyglet==2.0.18
PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.7
PyQt5==5.15.11
setuptools==75.6.0
wheel==0.45.1
